Extracted `mdbook watch` to separate file/module
